Added basic button interaction (set block, save)
