--- Assets/Scripts/Level/demoCaller.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23e2c2a0e808e364c85cf4994fe3a12b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
