+++ Assets/Scripts/LevelEditor/LevelEditor.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 /// <summary>
 /// Checks if Mous is clicking on a block on the grid and places new block if necessary
@@ -14,28 +15,29 @@ public class LevelEditor : MonoBehaviour
 
     private Block_Data currentBlockData = new EmptyBlock_Data();
 
+    private Level_Data data;
+
     #region Unity
     private void Awake()
     {
         Gamemaster.Instance.Register(this);
     }
 
+    void Start()
+    {
+        Gamemaster.Instance.GetLevel().CreateNewLevel(20, 20, "testLevel");
+        data = Gamemaster.Instance.GetLevel().GetLevelData();
+    }
+
     private void Update()
     {
-        //toDo: Replace simple KeyInput
-        for (int i = 0; i < 4; i++)
+        if (Input.GetMouseButton(0))
         {
-            if (Input.GetKeyDown("" + (i + 1)))
+            Vector2Int blockCoord = editorInput.GetBlockMouseIsOn();
+            if (blockCoord.x >= 0 && blockCoord.y >= 0)
             {
-                currentBlockData = blockDatas[i];
-                break;
+                Gamemaster.Instance.GetLevel().PlaceBlock(blockCoord.x, blockCoord.y, currentBlockData);
             }
-        }
-
-        Vector2Int blockCoord = editorInput.GetBlockMouseIsOn();
-        if (blockCoord.x >= 0 && blockCoord.y >= 0)
-        {
-            Gamemaster.Instance.GetLevel().PlaceBlock(blockCoord.x, blockCoord.y, currentBlockData);
         }
     }
 
@@ -44,5 +46,33 @@ public class LevelEditor : MonoBehaviour
     public void SetCurrentBlock(Block_Data blockData)
     {
         currentBlockData = blockData;
+    }
+
+    public void SetBlockType(int blockType)
+    {
+        if (blockType >= 0 && blockType < 4)
+        {
+            currentBlockData = blockDatas[blockType];
+        }
+    }
+
+    public void SetLevelName(string levelName)
+    {
+        data.Name = levelName;
+    }
+
+    public void SetLevelName(InputField levelNameInputField)
+    {
+        data.Name = levelNameInputField.text;
+    }
+
+    public void SaveLevel()
+    {
+        LevelSaveLoad.Save(data, "TestLevels");
+    }
+
+    public void LoadLevel()
+    {
+        Gamemaster.Instance.GetLevel().LoadLevelFromFile(data.Name, "TestLevels");
     }
 }
